Fixed calculation for CPU usage
Calculation for CPU usage is fixed, however the calculation for the heaviest process is still off. The calculation for the heaviest process will never be exact due to the process jumping between difrent processors over time, the recorded value will be the usage the process had and where it was last scheduled to run. However there is still something that is wrong in the calculation.